Fixed SPI Read Bug
Previously, the SPI connection used GPIO pins 18, 23, 24, and 25 along with code
published by Ladyada (Adafruit) to manually run the SPI protocol in Python. This
did not work and only reported a value of 0 from the connected ADC (MCP3204)
regardless of the voltage input. The reason for this error appears to be
insufficient timing control through the python script.

I updated the test.py code to use the mcp3204.py library published by yoggy. This
library uses the spidev library to run the SPI connection and reads the MCP3204
output based on the requested ADC channel. This code worked enough to provide
reasonable output values in test.py.

Also had to change the hardware configuration:
  GPIO18 -> SPICLK
  GPIO23 -> SPIMISO
  GPIO24 -> SPIMISI
  GPIO25 -> SPICS1